+++ modules/Users/Exceptions.php
@@ -226,8 +226,8 @@ if ( UserStaffID()
 					&& $file !== 'default'
 					&& $file !== 'title' )
 				{
-					$can_use = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_USE'] );
-					$can_edit = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_EDIT'] );
+					$can_use = issetVal( $exceptions_RET[$file][1]['CAN_USE'] );
+					$can_edit = issetVal( $exceptions_RET[$file][1]['CAN_EDIT'] );
 
 					echo '<td class="align-right"><input type="checkbox" name="can_use[' .
 					str_replace( '.', '_', $file ) . ']" value="true"' .
@@ -262,8 +262,8 @@ if ( UserStaffID()
 							$file = 'Students/Student.php&category_id=' . $category['ID'];
 							$title = '&nbsp;&nbsp;&rsaquo; ' . ParseMLField( $category['TITLE'] );
 
-							$can_use = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_USE'] );
-							$can_edit = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_EDIT'] );
+							$can_use = issetVal( $exceptions_RET[$file][1]['CAN_USE'] );
+							$can_edit = issetVal( $exceptions_RET[$file][1]['CAN_EDIT'] );
 
 							echo '<tr><td class="align-right"><input type="checkbox" name="can_use[' .
 							str_replace( '.', '_', $file ) . ']" value="true"' .
@@ -304,8 +304,8 @@ if ( UserStaffID()
 							$file = 'Users/User.php&category_id=' . $category['ID'];
 							$title = '&nbsp;&nbsp;&rsaquo; ' . ParseMLField( $category['TITLE'] );
 
-							$can_use = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_USE'] );
-							$can_edit = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_EDIT'] );
+							$can_use = issetVal( $exceptions_RET[$file][1]['CAN_USE'] );
+							$can_edit = issetVal( $exceptions_RET[$file][1]['CAN_EDIT'] );
 
 							echo '<tr><td class="align-right"><input type="checkbox" name="can_use[' .
 							str_replace( '.', '_', $file ) . ']" value="true"' .
@@ -326,8 +326,8 @@ if ( UserStaffID()
 								$file = 'Users/User.php&category_id=1&user_profile';
 								$title = ' &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&rsaquo; ' . _( 'User Profile' );
 
-								$can_use = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_USE'] );
-								$can_edit = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_EDIT'] );
+								$can_use = issetVal( $exceptions_RET[$file][1]['CAN_USE'] );
+								$can_edit = issetVal( $exceptions_RET[$file][1]['CAN_EDIT'] );
 
 								echo '<tr><td class="align-right"><input type="checkbox" name="can_use[' .
 								str_replace( '.', '_', $file ) . ']" value="true"' .
@@ -345,8 +345,8 @@ if ( UserStaffID()
 								$file = 'Users/User.php&category_id=1&schools';
 								$title = ' &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&rsaquo; ' . _( 'Schools' );
 
-								$can_use = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_USE'] );
-								$can_edit = ( empty( $exceptions_RET[$file] ) ? '' : $exceptions_RET[$file][1]['CAN_EDIT'] );
+								$can_use = issetVal( $exceptions_RET[$file][1]['CAN_USE'] );
+								$can_edit = issetVal( $exceptions_RET[$file][1]['CAN_EDIT'] );
 
 								echo '<tr><td class="align-right"><input type="checkbox" name="can_use[' .
 								str_replace( '.', '_', $file ) . ']" value="true"' .